Update to use smart ideas from talk.
This commit brings in some of the things I'd come up with prepping for
my HoustonJS talk on 3D rasterization.

This includes:

* Smarter OBJ loader
* Better bounds checking on the Bitmap

I've also gone ahead and added:

* Lambda-based depth function testing
* Depth-mask to check whether or not we're writing to the depth layer.

This also updates all the examples to reflect the new changes.

Lastly, I've added a picking callback which can be used for debugging.